requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras
#
alembic==1.13.2
    # via dagster
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via dagster
coloredlogs==14.0
    # via dagster
croniter==3.0.3
    # via dagster
dagster==1.7.16
    # via sgc-dagster (pyproject.toml)
dagster-pipes==1.7.16
    # via dagster
docstring-parser==0.16
    # via dagster
filelock==3.15.4
    # via dagster
fsspec==2024.6.1
    # via universal-pathlib
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.64.1
    # via
    #   dagster
    #   grpcio-health-checking
grpcio-health-checking==1.62.2
    # via dagster
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via sgc-dagster (pyproject.toml)
humanfriendly==10.0
    # via coloredlogs
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via dagster
mako==1.3.5
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
numpy==2.0.1
    # via pandas
packaging==24.1
    # via dagster
pandas==2.2.2
    # via sgc-dagster (pyproject.toml)
pendulum==3.0.0
    # via dagster
protobuf==4.25.4
    # via
    #   dagster
    #   grpcio-health-checking
pydantic==2.8.2
    # via dagster
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   croniter
    #   pandas
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via dagster
pytz==2024.1
    # via
    #   croniter
    #   dagster
    #   pandas
pyyaml==6.0.1
    # via dagster
requests==2.32.3
    # via dagster
rich==13.7.1
    # via dagster
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.31
    # via
    #   alembic
    #   dagster
structlog==24.4.0
    # via dagster
tabulate==0.9.0
    # via dagster
time-machine==2.14.2
    # via pendulum
tomli==2.0.1
    # via dagster
toposort==1.10
    # via dagster
tqdm==4.66.5
    # via dagster
typing-extensions==4.12.2
    # via
    #   alembic
    #   dagster
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.1
    # via
    #   pandas
    #   pendulum
universal-pathlib==0.2.2
    # via dagster
urllib3==2.2.2
    # via requests
watchdog==4.0.1
    # via dagster

# The following packages are considered to be unsafe in a requirements file:
# setuptools
